[BROOKLYN-183] Move Apache Felix code into specific jar
Move code that manages the Apache Felix framework into
brooklyn-rt-felix, with support code in brooklyn-rt-osgi.
Don't import packages from brooklyn-rt-felix in brooklyn-core.

Also move the OSGi test resources into brooklyn-rt-osgi.
