+++ utils/rt-felix/src/main/java/org/apache/brooklyn/rt/felix/ManifestHelper.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.brooklyn.util.core.osgi;
+package org.apache.brooklyn.rt.felix;
 
 import java.io.IOException;
 import java.io.InputStream;
