+++ utils/rt-felix/src/test/java/org/apache/brooklyn/rt/felix/EmbeddedFelixFrameworkTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.brooklyn.util.core.osgi;
+package org.apache.brooklyn.rt.felix;
 
 import java.io.File;
 import java.io.IOException;
@@ -23,13 +23,12 @@ import java.util.Enumeration;
 import java.util.List;
 import java.util.jar.JarInputStream;
 
-import org.apache.brooklyn.core.mgmt.osgi.OsgiStandaloneTest;
-import org.apache.brooklyn.core.mgmt.osgi.OsgiTestResources;
 import org.apache.brooklyn.test.support.TestResourceUnavailableException;
 import org.apache.brooklyn.util.collections.MutableSet;
-import org.apache.brooklyn.util.core.osgi.ManifestHelper;
 import org.apache.brooklyn.util.os.Os;
+import org.apache.brooklyn.util.osgi.OsgiTestResources;
 import org.apache.brooklyn.util.stream.Streams;
+import org.apache.commons.io.FileUtils;
 import org.osgi.framework.BundleException;
 import org.osgi.framework.launch.Framework;
 import org.slf4j.Logger;
@@ -37,6 +36,7 @@ import org.slf4j.LoggerFactory;
 import org.testng.Assert;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
 
 /**
  *
@@ -44,7 +44,7 @@ import org.testng.annotations.BeforeMethod;
  */
 public class EmbeddedFelixFrameworkTest {
 
-    private static final Logger log = LoggerFactory.getLogger(OsgiStandaloneTest.class);
+    private static final Logger log = LoggerFactory.getLogger(EmbeddedFelixFrameworkTest.class);
 
     public static final String BROOKLYN_TEST_OSGI_ENTITIES_PATH = OsgiTestResources.BROOKLYN_TEST_OSGI_ENTITIES_PATH;
 
@@ -54,17 +54,26 @@ public class EmbeddedFelixFrameworkTest {
     @BeforeMethod(alwaysRun = true)
     public void setUp() throws Exception {
         storageTempDir = Os.newTempDir("osgi-standalone");
-        framework = Osgis.getFramework(storageTempDir.getAbsolutePath(), false);
+        framework = EmbeddedFelixFramework.newFrameworkStarted(storageTempDir.getAbsolutePath(), true, null);
     }
 
     @AfterMethod(alwaysRun = true)
     public void tearDown() throws BundleException, IOException, InterruptedException {
-        OsgiStandaloneTest.tearDownOsgiFramework(framework, storageTempDir);
+        tearDownOsgiFramework(framework, storageTempDir);
     }
 
-    @org.testng.annotations.Test
+    public static void tearDownOsgiFramework(Framework framework, File storageTempDir) throws BundleException, InterruptedException, IOException {
+        EmbeddedFelixFramework.stopFramework(framework);
+        framework = null;
+        if (storageTempDir != null) {
+            FileUtils.deleteDirectory(storageTempDir);
+            storageTempDir = null;
+        }
+    }
+
+    @Test
     public void testReadAManifest() throws Exception {
-        Enumeration<URL> manifests = getClass().getClassLoader().getResources("META-INF/MANIFEST.MF");
+        Enumeration<URL> manifests = this.getClass().getClassLoader().getResources("META-INF/MANIFEST.MF");
         log.info("Bundles and exported packages:");
         MutableSet<String> allPackages = MutableSet.of();
         while (manifests.hasMoreElements()) {
@@ -75,10 +84,10 @@ public class EmbeddedFelixFrameworkTest {
         }
         log.info("Total export package count: " + allPackages.size());
         Assert.assertTrue(allPackages.size() > 20, "did not find enough packages"); // probably much larger
-        Assert.assertTrue(allPackages.contains(Osgis.class.getPackage().getName()));
+        Assert.assertTrue(allPackages.contains(EmbeddedFelixFramework.class.getPackage().getName()));
     }
 
-    @org.testng.annotations.Test
+    @Test
     public void testReadKnownManifest() throws Exception {
         TestResourceUnavailableException.throwIfResourceUnavailable(getClass(), BROOKLYN_TEST_OSGI_ENTITIES_PATH);
         InputStream in = this.getClass().getResourceAsStream(BROOKLYN_TEST_OSGI_ENTITIES_PATH);
