+++ utils/rt-osgi/src/test/resources/brooklyn/osgi/brooklyn-test-osgi-entities.jar
