+++ utils/rt-osgi/src/test/resources/dependencies/osgi/README.md
