+++ utils/rt-osgi/src/test/resources/dependencies/osgi/entities/pom.xml
