+++ utils/rt-osgi/src/test/resources/dependencies/osgi/entities/src/main/resources/org/apache/brooklyn/test/osgi/entities/icon.gif
