+++ utils/rt-osgi/src/test/resources/dependencies/osgi/more-entities-v1/pom.xml
