+++ utils/rt-osgi/src/test/resources/dependencies/osgi/more-entities-v2-evil-twin/pom.xml
